chatbot_executor_service/requirements.txt
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==45.0.3
dnspython==2.7.0
et_xmlfile==2.0.0
eventlet==0.39.0
Flask==3.1.0
Flask-SocketIO==5.5.1
gevent==24.11.1
gevent-websocket==0.10.1
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.67.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.3
oauthlib==3.2.2
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
paramiko==3.5.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyMySQL==1.1.1
PyNaCl==1.5.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.38
sshtunnel==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
XlsxWriter==3.2.2
zope.event==5.0
zope.interface==7.2
